Automatically cancel child jobs if main snakemake job in SLURM queue exits.
If snakemake is run in a SLURM job context, it can happen that the
snakemake job is cancelled, but the child jobs continue to run in the
queue and snakemake effectively loses track of those jobs.
To ensure that all child jobs are removed when the main snakemake job
is cancelled, a clean-up job is submitted to the queue that waits for
the main snakemake job to finish and then kills all remaining child jobs.
Child jobs are identified by their name which is equal to the <run_uuid>.